+++ config/Autoloader.php
@@ -1,7 +1,13 @@
 <?php
 
+/**
+ * This file contains Autoloader class
+ */
 namespace config;
 
+/**
+ * Class for autoloader of class 
+ */
 class Autoloader 
 {
     private static $_directories = [
@@ -12,15 +18,26 @@ class Autoloader
         'view/backView/',
         'config/'];
 
+    /**
+     * __construct Init 
+     * 
+     * @return void
+     */
     public function __construct()
     {
         spl_autoload_register([__CLASS__, 'classLoader']);
         
     }
 
+    /**
+     * Test class
+     * 
+     * @param string $className Name of class
+     * 
+     * @return void
+     */
     public static function classLoader($className)
     {
-        
         foreach (self::$_directories as $directory) {
             $nameSpace = explode('\\', $className);
