+++ config/config.php
@@ -1,6 +1,8 @@
 <?php
 
-//CONSTANTS
+/**
+ * This file contains PARAMETERS AND MESSAGES
+ */
 
 //Database config
 define('HOST', 'mysql:host=localhost');
