+++ controller/Controller.php
@@ -1,19 +1,36 @@
 <?php
 
+/**
+ * This file contains Controller class
+ */
 namespace controller;
 use Twig;
+use Twig_Extensions_Extension_Text;
 
+/**
+ * Class for init Twig, send email and control session
+ */
 class Controller
 {
     protected $twig;
     protected $user;
     protected $msg;
 
+    /**
+     * Getter $msg
+     * 
+     * @return string
+     */
     public function msg()
     {
         return $this->msg;
     }
 
+    /**
+     * __construct Init object
+     * 
+     * @return void
+     */
     public function __construct()
     {
         $var = new \config\GlobalVar;
@@ -25,6 +42,11 @@ class Controller
         }
     }
 
+    /**
+     * Init twig
+     * 
+     * @return void
+     */
     public function twigInit()
     {
         $loader = new Twig\Loader\FilesystemLoader('view');
@@ -33,6 +55,14 @@ class Controller
         ); 
     }
 
+    /**
+     * Method for send emails
+     * 
+     * @param array  $form  Input data in form
+     * @param string $email Email format destination email
+     * 
+     * @return void
+     */
     public function sendMessage(array $form, $email)
     {
         $var = new \config\GlobalVar;
@@ -53,6 +83,13 @@ class Controller
         $this->msg = $msg;
     }
 
+    /**
+     * Control session
+     * 
+     * @param string $name Index of session
+     * 
+     * @return bool
+     */
     public function sessionControl($name)
     {
         $var = new \config\GlobalVar;
@@ -61,5 +98,25 @@ class Controller
             return true;
         }
         return false;
+    }
+
+    /**
+     * Display error page
+     * 
+     * @param string $exception Error message define in config.php
+     * 
+     * @return void
+     */
+    public function errorView($exception)
+    {
+        $this->twigInit();
+        $this->twig->addExtension(new Twig\Extension\DebugExtension); //think to delete this line
+        $this->twig->addExtension(new Twig_Extensions_Extension_Text()); 
+
+        echo $this->twig->render(
+            'errorView.twig', array(
+                'exception' => $exception 
+            )
+        );
     }
 }
