+++ controller/backCommentController.php
@@ -1,11 +1,24 @@
 <?php
 
+/**
+ * This file contains BackCommentController class
+ */
 namespace controller;
+
+use Exception;
 use Twig;
 use Twig_Extensions_Extension_Text;
 
+/**
+ * Class for get comments data and send it back to views
+ */
 class BackCommentController extends BackController
 {
+    /**
+     * Retrieves and send data to valid comment page
+     * 
+     * @return void
+     */
     public function validComment()
     {
         $commentManager = new \model\CommentManager;
@@ -22,6 +35,13 @@ class BackCommentController extends BackController
         );
     }
 
+    /**
+     * Update comment - Comment change to valid
+     * 
+     * @param int $commentId Id of comment
+     * 
+     * @return Number of affected lines
+     */
     public function updateComment($commentId)
     {
         $commentManager = new \model\CommentManager;
@@ -29,10 +49,22 @@ class BackCommentController extends BackController
         return $affectedLine;
     }
 
+    /**
+     * Retrieves and send comment data to delete comment page
+     * 
+     * @param int $commentId Id of comment
+     * 
+     * @return void
+     */
     public function deleteCommentView($commentId)
     {
+        $var = new \config\GlobalVar;
+
         $commentManager = new \model\CommentManager;
         $data = $commentManager -> getComment($commentId);
+        if ($data == false) {
+            throw new \Exception(PAGE_NOT_EXIST);
+        }
         $comment = new \model\Comment($data);
 
         $this->twigInit();
@@ -44,9 +76,14 @@ class BackCommentController extends BackController
                 'comment' => $comment 
             )
         );
-        $_SESSION['comment'] = $comment;
+        $var->setSession('comment', $comment);
     }
 
+    /**
+     * Delete comment
+     * 
+     * @return void
+     */
     public function deleteComment()
     {
         $var = new \config\GlobalVar;
@@ -61,10 +98,19 @@ class BackCommentController extends BackController
             $affectedLine = $commentManager -> deleteComment(
                 $var->session('comment')->commentId()
             );
+
             $this -> deleteCommentMail($affectedLine, $addMsgEmail);
         }   
     }
 
+    /**
+     * Send an email to visitor who write comment that will be delete
+     * 
+     * @param int    $affectedLine Number of affected line when delete comment
+     * @param string $addMsgEmail  Message send to a visitor
+     * 
+     * @return void
+     */
     public function deleteCommentMail($affectedLine, $addMsgEmail)
     {
         $var = new \config\GlobalVar;
@@ -93,6 +139,13 @@ class BackCommentController extends BackController
         throw new \Exception(COMMENT_NO_EXIST);   
     }
 
+    /**
+     * Retrieves and send comment data to list comment page
+     * 
+     * @param string $try Try of comment request results (date, author, post)
+     * 
+     * @return void
+     */
     public function listComments($try=null)
     {
         $commentManager = new \model\CommentManager;
@@ -109,6 +162,11 @@ class BackCommentController extends BackController
         );
     }
 
+    /**
+     * Valid delete comment
+     * 
+     * @return void
+     */
     public function validDeleteComment()
     {
         $var = new \config\GlobalVar;
@@ -137,6 +195,11 @@ class BackCommentController extends BackController
         return;
     }
 
+    /**
+     * Action buttons on list comment page
+     * 
+     * @return void
+     */
     public function listCommentsAction()
     {
         $var = new \config\GlobalVar;
