+++ controller/backController.php
@@ -1,11 +1,22 @@
 <?php
 
+/**
+ * This file contains BackController class
+ */
 namespace controller;
 use Twig;
 use Twig_Extensions_Extension_Text;
 
+/**
+ * Class for display back home page
+ */
 class BackController extends Controller
 {
+    /**
+     * Display back home page
+     * 
+     * @return void
+     */
     public function backHomePage()
     {
         $infos = $this -> backHomePostBadge();
@@ -31,6 +42,11 @@ class BackController extends Controller
         );
     }
 
+    /**
+     * Retrieves data for back home page badges
+     * 
+     * @return array $infos Infos of badges
+     */
     public function backHomePostBadge()
     {
         $postManager = new \model\PostManager;
