+++ controller/backPImgController.php
@@ -1,11 +1,24 @@
 <?php
 
+/**
+ * This file contains BackPImgController class
+ */
 namespace controller;
 use Twig;
 use Twig_Extensions_Extension_Text;
 
+/**
+ * Class for manage post image
+ */
 class BackPImgController extends BackPostController
 {
+    /**
+     * Test and manage upload post image
+     * 
+     * @param string $imgPost Name of picture post if exist
+     * 
+     * @return string Path of picture post
+     */
     public function uploadFile($imgPost=null)
     {  
         $var = new \config\GlobalVar;
@@ -35,6 +48,14 @@ class BackPImgController extends BackPostController
         }      
     }
 
+    /**
+     * Move file for save upload file in definity folder
+     * 
+     * @param string $imgPost   Name of picture post
+     * @param string $directory Destination directory of picture post
+     * 
+     * @return void 
+     */
     public function moveFile($imgPost,$directory)
     {
         move_uploaded_file(
@@ -43,6 +64,12 @@ class BackPImgController extends BackPostController
         );
     }
 
+    /**
+     * Copy picture in definitive folder
+     * 
+     * @param string $imgPost   Name of picture post
+     * @param string $directory Definitive directory of picture post
+     */
     public function renameFile($imgPost, $directory)
     {
         rename(
@@ -52,6 +79,12 @@ class BackPImgController extends BackPostController
         );
     }
 
+    /**
+     * Rename picture post with a unique name with time() fonction
+     * 
+     * @return mixed string if addPost or notPublished
+     *               null if preview
+     */
     public function managePostImage() 
     {
         $var = new \config\GlobalVar;
@@ -76,6 +109,11 @@ class BackPImgController extends BackPostController
         return null;           
     }
 
+    /**
+     * Change picture post and delete old picture
+     * 
+     * @return void
+     */
     public function imgChange()
     {
         $var = new \config\GlobalVar;
