+++ controller/backUserController.php
@@ -1,11 +1,26 @@
 <?php
 
+/**
+ * This file contains BackUserController class
+ */
 namespace controller;
 use Twig;
 use Twig_Extensions_Extension_Text;
 
+/**
+ * Class for get users data and send it back to views
+ */
 class BackUserController extends BackController
 {
+
+    /**
+     * Retrives and send data to addUser page
+     * 
+     * @param array $form Null if add a new user
+     *                    Array with user data if update a user
+     * 
+     * @return void 
+     */
     public function addUserView($form=null)
     {
         $var = new \config\GlobalVar;
@@ -35,6 +50,13 @@ class BackUserController extends BackController
         );
     }
 
+    /**
+     * Create a user object for add a user in database
+     * 
+     * @param array $form Array with inputs adduser form
+     * 
+     * @return int Number of affected lines in database 
+     */
     public function addUser($form)
     {
         $userManager = new \model\UserManager;
@@ -43,6 +65,11 @@ class BackUserController extends BackController
         return $userManager -> addUser($user); 
     }
 
+    /**
+     * Test inputs in adduser form on adduser page
+     * 
+     * @return void
+     */
     public function testAdduser()
     {
         $var = new \config\GlobalVar;
@@ -79,6 +106,11 @@ class BackUserController extends BackController
         exit();   
     }
 
+    /**
+     * Test inputs in adduser form on adduser page when update a user
+     * 
+     * @return int Number of affected lines 
+     */
     public function testUpdateUser()
     {
         $frontController = new \controller\FrontController;
@@ -109,6 +141,11 @@ class BackUserController extends BackController
         return $affectedLine;
     }
 
+    /**
+     * Retrieves and send data to list users page
+     * 
+     * @return void
+     */
     public function listUsers()
     {
         $userManager = new \model\UserManager;
@@ -123,27 +160,45 @@ class BackUserController extends BackController
         );
     }
 
-    public function getUpdateUser($id)
+    /**
+     * Retrieves data user for update a user
+     * 
+     * @param int $userId 
+     * 
+     * @return array User data
+     */
+    public function getUpdateUser($userId)
     {
         $userManager = new \model\UserManager;
-        $data = $userManager -> getUser((int)$id);
+        $data = $userManager -> getUser((int)$userId);
         if ($data != false) {
             return $data;
         } else {
             throw new \Exception(USER_NO_EXIST);
-        }
-          
+        }    
     }
 
+    /**
+     * Retrieves inputs in addUser form on addUser page for update a user
+     * 
+     * @param array $form Inputs in adduser form
+     * 
+     * @return int Number of affected lines
+     */
     public function updateUser($form)
     {
         $userManager = new \model\UserManager;
-
         $user = new \model\User($form);
-        
         return $userManager -> updateUser($user);
     }
 
+    /**
+     * Retrieves and send data user to delete user page
+     * 
+     * @param int $userId Id of user
+     * 
+     * @return void
+     */
     public function deleteUserView($userId)
     {
         $userManager = new \model\UserManager;
@@ -163,10 +218,16 @@ class BackUserController extends BackController
                 'deleteUser' => $deleteUser,
                 'posts' => $posts 
             )
-        );
-        
+        ); 
     }
 
+    /**
+     * Delete user 
+     * 
+     * @param int $userId Id of user
+     * 
+     * @return void 
+     */
     public function deleteUser($userId)
     {
         $var = new \config\GlobalVar;
@@ -186,6 +247,15 @@ class BackUserController extends BackController
         return;
     }
 
+    /**
+     * Retrieves and send data user to profil page
+     * 
+     * @param int   $userId Id of user
+     * @param mixed $update Int $update=1 for modify profil inputs
+     *                      else Null inputs are on ReadOnly
+     * 
+     * @return void
+     */
     public function profilView($userId, $update=null)
     {
         $var = new \config\GlobalVar;
@@ -222,6 +292,11 @@ class BackUserController extends BackController
         );
     }
 
+    /**
+     * Test inputs in profil form
+     * 
+     * @return void
+     */
     public function dataInputProfil()
     {
         $var = new \config\GlobalVar;
@@ -273,6 +348,13 @@ class BackUserController extends BackController
         exit();
     }
 
+    /**
+     * Upload profil picture in profil form
+     * 
+     * @param string $picture Picture name
+     * 
+     * @return string Picture full name - path of file
+     */
     public function uploadProfilPicture($picture)
     {
         $this -> deleteOldProfilPicture();
@@ -299,6 +381,11 @@ class BackUserController extends BackController
         }
     }
 
+    /**
+     * Delete old profil picture when change
+     * 
+     * @return void
+     */
     public function deleteOldProfilPicture() 
     {
         $var = new \config\GlobalVar;
